Revised as per Paul Sefton comments
Slide 31 – there are 6 alleluias (2 groups of 3) before and after the
verse. Not important though.
Slide 59 – “until you come again” is not repeated in the St. Gregory
Mass setting.
Slide 61 – in this Mass setting, there are 6 amens, again in 2 groups of
3.
Slide 65 – the last line should be “Grant us peace” – no your.